move TreeNode
