+++ airbyte-db/db-lib/src/main/java/io/airbyte/db/mongodb/TreeNode.java
@@ -2,7 +2,7 @@
  * Copyright (c) 2022 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.protocol.models;
+package io.airbyte.db.mongodb;
 
 import java.util.LinkedList;
 import java.util.List;
@@ -13,13 +13,13 @@ public class TreeNode<T> {
   private TreeNode<T> parent;
   private List<TreeNode<T>> children;
 
-  public TreeNode(T data) {
+  public TreeNode(final T data) {
     this.data = data;
     this.children = new LinkedList<>();
   }
 
-  public TreeNode<T> addChild(T child) {
-    TreeNode<T> childNode = new TreeNode<T>(child);
+  public TreeNode<T> addChild(final T child) {
+    final TreeNode<T> childNode = new TreeNode<T>(child);
     childNode.parent = this;
     this.children.add(childNode);
     return childNode;
@@ -33,7 +33,7 @@ public class TreeNode<T> {
     return data;
   }
 
-  public void setData(T data) {
+  public void setData(final T data) {
     this.data = data;
   }
 
@@ -41,7 +41,7 @@ public class TreeNode<T> {
     return parent;
   }
 
-  public void setParent(TreeNode<T> parent) {
+  public void setParent(final TreeNode<T> parent) {
     this.parent = parent;
   }
 
@@ -49,7 +49,7 @@ public class TreeNode<T> {
     return children;
   }
 
-  public void setChildren(List<TreeNode<T>> children) {
+  public void setChildren(final List<TreeNode<T>> children) {
     this.children = children;
   }
